src/database/requirements.txt
mysql-connector-python==8.0.33
python-dotenv==0.21.1
boto3==1.26.0
botocore==1.29.0
